requirements.txt
markdown-it-py[linkify,plugins]
feedgenerator
rich
Pygments